Move AppRoutes.tsx to router/router.tsx
