+++ frontend/src/index.tsx
@@ -5,7 +5,7 @@ import '~/global.scss';
 import { GlobalContextProvider } from '~/context/GlobalContextProvider';
 import { reportWebVitals } from '~/reportWebVitals';
 import { RouterProvider } from 'react-router-dom';
-import { router } from '~/AppRoutes';
+import { router } from '~/router/router';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
